move static to web, and remove unused code
