--- pages/reader/article.html
@@ -1,17 +1,0 @@
-<link rel="stylesheet" href="/css/article.css" />
-<link rel="stylesheet" href="/css/markdown.css" />
-<div class="container">
-  <div class="article-header-links">
-    <a href="/" class="back-link">← Volver al PyE Times</a>
-    {{ edit_link }}
-  </div>
-
-  <article class="article-header">
-    <h1 class="headline">{{ title }}</h1>
-    <h2 class="subheadline">{{ excerpt }}</h2>
-    <div class="byline">Por {{ author_name }}</div>
-    <div class="date-info">Publicado: {{ published_at }}</div>
-  </article>
-
-  <div class="article-content">{{ content }}</div>
-</div>
