--- pages/reader/article_preview.html
@@ -1,6 +1,0 @@
-<article class="article">
-  <h3 class="article-headline">
-    <a href="/articles/{{ slug }}">{{ title }}</a>
-  </h3>
-  <p class="article-excerpt">{{ excerpt }}</p>
-</article>
