--- pages/reader/main_story.html
@@ -1,8 +1,0 @@
-<main class="main-story">
-  <h2 class="main-headline">
-    <a href="/articles/{{ slug }}">{{ headline }}</a>
-  </h2>
-  <p class="byline">Por {{ author_name }}</p>
-  <p class="date">{{ date }}</p>
-  <p class="lead-text">{{ lead_text }}</p>
-</main>
