--- pages/reader/section.html
@@ -1,4 +1,0 @@
-<section class="section">
-  <h2 class="section-header">{{ title }}</h2>
-  {{ articles }}
-</section>
