--- static/css/article.css
@@ -1,302 +1,0 @@
-:root {
-  /* Colores principales */
-  --color-primary: #d32f2f;
-  --color-primary-dark: #b71c1c;
-  --color-black: #000;
-  --color-white: #fff;
-
-  /* Grises */
-  --color-text-primary: #333;
-  --color-text-secondary: #666;
-  --color-text-tertiary: #777;
-  --color-text-muted: #999;
-  --color-text-warning: #856404;
-  --color-text-light: #6c757d;
-
-  /* Fondos */
-  --color-bg-dark: #343a40;
-  --color-bg-darker: #2d2d2d;
-  --color-bg-darkest: #1a1a1a;
-  --color-bg-light-gray: #e9ecef;
-  --color-bg-warning: #fff3cd;
-
-  /* Bordes */
-  --color-border-primary: #000;
-  --color-border-secondary: #ddd;
-  --color-border-light: #eee;
-  --color-border-warning: #ffc107;
-  --color-border-gray: #adb5bd;
-
-  /* Enlaces */
-  --color-link: #007bff;
-  --color-link-hover: #d32f2f;
-
-  /* Efectos especiales */
-  --color-primary-light: #ff6b6b;
-}
-
-.tagline {
-  font-size: 0.9em;
-  font-style: italic;
-  color: var(--color-text-secondary);
-  margin-bottom: 10px;
-}
-
-.investigation-banner {
-  background: linear-gradient(135deg, var(--color-primary-dark) 0%, var(--color-primary) 100%);
-  color: var(--color-white);
-  text-align: center;
-  padding: 12px;
-  font-weight: bold;
-  font-size: 1em;
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.3);
-}
-
-.container {
-  max-width: 1000px;
-  margin: 0 auto;
-  padding: 0px 20px;
-}
-
-.article-header {
-  margin-bottom: 50px;
-  border-bottom: 2px solid var(--color-border-secondary);
-  padding-bottom: 40px;
-}
-
-.category {
-  color: var(--color-primary);
-  font-size: 0.9em;
-  font-weight: bold;
-  text-transform: uppercase;
-  letter-spacing: 1px;
-  margin-bottom: 10px;
-}
-
-.headline {
-  font-size: 2.8em;
-  font-weight: bold;
-  line-height: 1.1;
-  margin-bottom: 15px;
-  color: var(--color-black);
-}
-
-.subheadline {
-  font-size: 1.4em;
-  color: var(--color-text-secondary);
-  font-weight: normal;
-  line-height: 1.3;
-  margin-bottom: 20px;
-}
-
-.byline {
-  font-size: 0.95em;
-  margin-bottom: 10px;
-  color: var(--color-text-secondary);
-}
-
-.date-info {
-  font-size: 0.9em;
-  color: var(--color-text-muted);
-}
-
-.warning-box {
-  background: var(--color-bg-warning);
-  border: 2px solid var(--color-warning);
-  border-radius: 8px;
-  padding: 20px;
-  margin: 25px 0;
-  position: relative;
-}
-
-.warning-box::before {
-  content: "⚠️";
-  font-size: 1.5em;
-  position: absolute;
-  top: 15px;
-  left: 15px;
-}
-
-.warning-text {
-  margin-left: 40px;
-  font-size: 0.95em;
-  color: var(--color-text-warning);
-  font-weight: 500;
-}
-
-.dropcap {
-  float: left;
-  font-size: 4em;
-  line-height: 1;
-  margin: 0 8px 0 0;
-  font-weight: bold;
-  color: var(--color-primary);
-}
-
-.quote-box {
-  background: var(--color-bg-secondary);
-  border-left: 4px solid var(--color-primary);
-  padding: 20px;
-  margin: 25px 0;
-  font-style: italic;
-  font-size: 1.05em;
-}
-
-.quote-attribution {
-  text-align: right;
-  margin-top: 10px;
-  font-size: 0.9em;
-  color: var(--color-text-secondary);
-  font-style: normal;
-}
-
-.evidence-section {
-  background: var(--color-bg-tertiary);
-  border: 1px solid var(--color-border-secondary);
-  border-radius: 8px;
-  padding: 25px;
-  margin: 30px 0;
-}
-
-.evidence-title {
-  font-size: 1.3em;
-  font-weight: bold;
-  margin-bottom: 15px;
-  color: var(--color-primary);
-}
-
-.evidence-item {
-  margin-bottom: 15px;
-  padding-left: 20px;
-  position: relative;
-}
-
-.evidence-item::before {
-  content: "📊";
-  position: absolute;
-  left: 0;
-}
-
-.screenshot-placeholder {
-  background: var(--color-bg-light-gray);
-  border: 2px dashed var(--color-border-gray);
-  border-radius: 8px;
-  padding: 40px;
-  text-align: center;
-  margin: 20px 0;
-  color: var(--color-text-light);
-}
-
-.staff-section {
-  background: linear-gradient(135deg, var(--color-bg-darkest) 0%, var(--color-bg-darker) 100%);
-  color: var(--color-white);
-  padding: 30px;
-  border-radius: 10px;
-  margin: 30px 0;
-  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.3);
-}
-
-.staff-title {
-  font-size: 1.5em;
-  font-weight: bold;
-  margin-bottom: 20px;
-  color: var(--color-primary-light);
-}
-
-.network-diagram {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 20px;
-  margin: 20px 0;
-}
-
-.network-node {
-  background: var(--color-bg-dark);
-  padding: 15px;
-  border-radius: 8px;
-  text-align: center;
-  border: 2px solid var(--color-primary-light);
-  position: relative;
-}
-
-.network-node::after {
-  content: "🔗";
-  position: absolute;
-  top: -10px;
-  right: -10px;
-  background: var(--color-primary-light);
-  border-radius: 50%;
-  padding: 5px;
-  font-size: 0.8em;
-}
-
-.conclusion-section {
-  background: var(--color-bg-primary);
-  border: 3px solid var(--color-primary);
-  border-radius: 10px;
-  padding: 30px;
-  margin: 40px 0;
-  box-shadow: 0 5px 20px rgba(211, 47, 47, 0.1);
-}
-
-.conclusion-title {
-  font-size: 1.6em;
-  font-weight: bold;
-  color: var(--color-primary);
-  margin-bottom: 20px;
-  text-align: center;
-}
-
-.footer-note {
-  background: var(--color-bg-secondary);
-  border-top: 3px solid var(--color-black);
-  padding: 20px;
-  margin-top: 40px;
-  font-size: 0.9em;
-  color: var(--color-text-secondary);
-  text-align: center;
-}
-
-.footer-note strong {
-  color: var(--color-primary);
-}
-
-@media (max-width: 768px) {
-  .headline {
-    font-size: 2.2em;
-  }
-
-  .network-diagram {
-    grid-template-columns: 1fr;
-  }
-
-  .container {
-    padding: 20px 15px;
-  }
-}
-
-.back-link {
-  display: inline-block;
-  margin-bottom: 20px;
-  color: var(--color-primary);
-  text-decoration: none;
-  font-size: 0.9em;
-}
-
-.back-link:hover {
-  text-decoration: underline;
-}
-
-.article-header-links {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.edit-link {
-  color: var(--color-link);
-  text-decoration: none;
-  font-size: 0.9em;
-}
